+++ Table/Selection/SelectionButtonBuilder.php
@@ -2,6 +2,11 @@
 
 namespace JGM\TableBundle\Table\Selection;
 
+use JGM\TableBundle\Table\AccessValidation\AccessValidatorFactory;
+use JGM\TableBundle\Table\Selection\Button\SubmitButton;
+use Symfony\Component\DependencyInjection\ContainerInterface;
+use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
+
 /**
  * Builder for buttons, used for submitting selected 
  * rows of the table.
@@ -12,18 +17,34 @@ namespace JGM\TableBundle\Table\Selection;
 class SelectionButtonBuilder
 {
 	/**
+	 * @var AuthorizationCheckerInterface
+	 */
+	private $authorizationChecker;
+
+	/**
 	 * @var array
 	 */
 	private $buttons;
 	
-	public function __construct()
+	public function __construct(ContainerInterface $container)
 	{
+		$this->authorizationChecker = $container->get('security.authorization_checker');
 		$this->buttons = array();
 	}
 	
 	public function add($name, array $options = array())
 	{
-		$this->buttons[$name] = new Button\SubmitButton($name, $options);
+		if(array_key_exists('access', $options))
+		{
+			if($this->isAccessGranted($options['access']) === false)
+			{
+				return $this;
+			}
+			
+			unset($options['access']);
+		}
+		
+		$this->buttons[$name] = new SubmitButton($name, $options);
 		
 		return $this;
 	}
@@ -31,5 +52,16 @@ class SelectionButtonBuilder
 	public function getButtons()
 	{
 		return $this->buttons;
+	}
+	
+	private function isAccessGranted($access)
+	{
+		$validator = AccessValidatorFactory::getValidator($access);
+		if($validator === null)
+		{
+			return false;
+		}
+		
+		return $validator->isAccessGranted($this->authorizationChecker);
 	}
 }